requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
colorama==0.4.6
colorlog==6.9.0
croniter==6.0.0
discord-py-interactions==5.15.0
discord-typings==0.9.0
emoji==2.15.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tomli==2.2.1
typing_extensions==4.15.0
yarl==1.20.1
